Resolve code style/format issues.
