+++ server/src/tests/namedtuple.test.ts
@@ -1,5 +1,5 @@
 /*
-* semanticAnalyzer.test.ts
+* namedtuple.test.ts
 * Copyright (c) Microsoft Corporation.
 * Licensed under the MIT license.
 * Author: Eric Traut
@@ -19,9 +19,9 @@ test('NamedTuple', () => {
 });
 
 test('NamedTuple field name error', () => {
-  let [analysisResult] = TestUtils.typeAnalyzeSampleFiles(['namedtuple2.py']);
+    let [analysisResult] = TestUtils.typeAnalyzeSampleFiles(['namedtuple2.py']);
 
-  assert.equal(analysisResult.errors.length, 1);
-  assert.ok(/start with _/.test(analysisResult.errors[0].message));
-  assert.equal(analysisResult.warnings.length, 0);
+    assert.equal(analysisResult.errors.length, 1);
+    assert.ok(/start with _/.test(analysisResult.errors[0].message));
+    assert.equal(analysisResult.warnings.length, 0);
 });
