+++ server/src/tests/samples/namedtuple2.py
@@ -1,4 +1,10 @@
 from typing import NamedTuple
 
+"""
+
+In Python 3.7, a NamedTuple field can not start with '_'
+
+"""
+
 class DataTuple2(NamedTuple):
-  _id: int
+    _id: int
